Enamel/Dentin-detection more error-tolerant
If we can't find a border, just assume it's at slice 500 and return the
slices from there.
This means that we'll have to redo-these detections manually, after
inspecting *all* of them, but that's not such a hassle :)

Additionally, all around a bit more error-tolerance built in, so we can
run the notebook from start to finish.
Intermediate results are loaded from disk, so it's simple to just delete
them to let them be regenerated...